Add files via upload
venv folder is the virtual environment to ignore.

Typically, in a real-world full-stack web application, you'd separate the front-end and back-end. However, Flask (the back-end framework) is designed to serve HTML templates directly from the server side, which is why it expects the HTML files to be inside a templates folder within the back-end structure. Since this is a portfolio project where the front-end and back-end are combined for simplicity, I'm keeping the HTML inside the back-end’s templates folder so that Flask can render it directly.

If you want to switch to a more decoupled approach later, you can serve the front-end from a separate front-end app and have the back-end (Flask) serve API endpoints.